### 2022-12-21 #566 @RWParsons
predictNMB: Clinical Prediction Models by Net Monetary Benefit
Submitting Author Name: Rex Parsons
Submitting Author Github Handle: @RWParsons
Other Package Authors Github handles: @robinblythe, @agbarnett
Repository:  https://github.com/RWParsons/predictNMB
Version submitted: 0.0.1
Submission type: Stats
Badge grade: silver
Version accepted: TBD
---

```
Package: predictNMB
Title: Evaluate Clinical Prediction Models by Net Monetary Benefit
Version: 0.0.1
    person("Rex", "Parsons", role = c("aut", "cre"),
           email = "rex.parsons94@gmail.com",
    person("Robin", "Blythe", role = "aut",
    person("Adrian", "Barnett", role = "aut",
    person("Susanna", "Cramb", role = "ctb",
    person("Steven", "McPhail", role = "ctb",
           comment = c(ORCID = "0000-0002-1463-662X"))
  )
Description: Estimates when and where a model-guided treatment strategy may outperform
    a treat-all or treat-none approach by Monte Carlo simulation and evaluation of the Net Monetary Benefit (NMB).
Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    assertthat,
    cutpointr,
    pmsampsize,
    tidyr
    covr,
    flextable,
    parallel,
    pbapply,
    vdiffr,
    withr

```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [x] A pre-submission inquiry has been approved in [issue#564](#564)

    - People considering the development of a clinical decision support system used to guide patient care between two possible interventions (or non-treatment) and wanting to know how likely it is that their model will be clinically useful compared to a treat-all or treat-none strategy. Performing simulations using hypothetical clinical prediction models of likely estimated performance may show that value is unlikely and deter researchers from model development efforts. These models are rarely used in clinical practice due to them often failing to improve patient outcomes compared to standard care. All evaluation is done in health economic terms, to orient the results towards value-based care.

    - This is the first implementation as far as I know. This package does involve a process to simulate clinical prediction models given a user's input of model discrimination (AUC) - the process to transform the AUC to a Cohen's D value is proposed here <http://dx.doi.org/10.5093/ejpalc2018a5> and we have a journal article currently under review which uses it for the same purpose as in this package.

    - Yes

    - Silver

    - Compliance with a good number of standards beyond those identified as minimally necessary (EA : 30 / 34; G : 39 / 68; Total : 69 / 102)
    - Generality: Has been developed to be useful for a wide range of clinical use cases - wherever a clinical prediction model may be used to guide the assignment of a patient to one of two possible treatments (or non-treatment). Functions have been made to be versatile to this wider range of use-cases (particularly `get_nmb_sampler`).

    - I have attempted to run `autotest` but I get an error when I do so when testing the example code. I get the error even when I remove the documentation that initially raises the error, so think there might be an issue with autotest (?)
