+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   detailed-example: detailed-example.html
   introduction-to-predictNMB: introduction-to-predictNMB.html
   summarising-results-with-predictNMB: summarising-results-with-predictNMB.html
-last_built: 2022-12-21T02:58Z
+last_built: 2022-12-21T04:35Z
 urls:
   reference: https://rwparsons.github.io/predictNMB/reference
   article: https://rwparsons.github.io/predictNMB/articles
