### 2022-12-23 #566 @annakrystalli
Hello again @RWParsons ,

Just a not that most editors are on a holiday break. We will pick up assigning a handling editor as soon as everyone is back in early January. Have a nice break! 🎄
