### 2023-01-04 08c9f66
add progress bar (with show_progress argument)
