### 2023-01-04 445b279
add show_progress example to intro vignette
