### 2023-01-04 2fef46d
test for correct class rather than output
for some reason, the printed progress bar from pbapply doesn't seem to register with expect_output()
