+++ tests/testthat/test-do_nmb_sim.R
@@ -10,15 +10,13 @@ test_that("do_nmb_sim() works", {
   expect_equal(nrow(out$df_result), nrow(stats::na.omit(out$df_result)))
   expect_equal(nrow(out$df_thresholds), nrow(stats::na.omit(out$df_thresholds)))
 
-  f <- function() {
-    do_nmb_sim(
-      sample_size = 100, n_sims = 10, n_valid = 1000, sim_auc = 0.7,
-      event_rate = 0.1,  fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
-      show_progress = TRUE
-    )
-  }
+  out <- do_nmb_sim(
+    sample_size = 100, n_sims = 10, n_valid = 1000, sim_auc = 0.7,
+    event_rate = 0.1,  fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
+    show_progress = TRUE
+  )
 
-  expect_output(f(), regexp = "100% elapsed", fixed = TRUE)
+  expect_s3_class(out, "predictNMBsim")
 })
 
 
@@ -261,15 +259,13 @@ test_that("do_nmb_sim() works in parallel", {
     skip()
   }
 
-  f <- function() {
-    do_nmb_sim(
-      sample_size = 200, n_sims = 100, n_valid = 1000, sim_auc = 0.7,
-      event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
-      cl = cl, cutpoint_methods = c("all", "none"), show_progress = TRUE
-    )
-  }
+  out_par_progress <- do_nmb_sim(
+    sample_size = 200, n_sims = 100, n_valid = 1000, sim_auc = 0.7,
+    event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
+    cl = cl, cutpoint_methods = c("all", "none"), show_progress = TRUE
+  )
 
-  expect_output(f(), regexp = "100% elapsed", fixed = TRUE)
+  expect_s3_class(out_par_progress, "predictNMBsim")
 
   parallel::stopCluster(cl)
 })
