### 2023-01-04 #566 @RWParsons
@annakrystalli  I just snuck in a quick change to the package to add progress bars which I had been meaning to do before submission but forgotted about until yesterday - I've updated the description in the OP of this issue and will re-run the bot to check the package. Hope this is all fine - thanks!
