### 2023-02-05 #566 @adamhsparks
Hi @RWParsons,
I'll be your handling editor for this submission.
This is my first stats package submission, so please bear with me, but I think we'll be alright. ☺️

I've performed the editor checks on your package and have a few items that should be addressed before the reviewers start their work.

---

##### Vignettes and documentation

- If you were to rename the "introduction-to-predictNMB.Rmd" to "predict_NMB.Rmd" it would follow a common schema in R packages, especially those that use {pkgdown} and you would have a "Get Started" menu item for the introductory vignette for users to easily find, _e.g._ <https://docs.ropensci.org/nasapower/articles/nasapower.html>.

##### Package integrity

- There should not be a `lazydata: true` field in the DESCRIPTION file, there is no data in the package

>   Omitted ‘LazyData’ from DESCRIPTION

- The dependency on R>= 3.5.0 should be set for .Rds objects used in testing

> NB: this package now depends on R (>= 3.5.0)
     WARNING: Added dependency on R >= 3.5.0 because serialized objects in
     serialize/load version 3 cannot be read in older versions of R.
     File(s) containing such objects:
       ‘predictNMB/tests/testthat/fixtures/predictNMBscreen_object.rds’
       ‘predictNMB/tests/testthat/fixtures/predictNMBsim_object.rds’

##### Tests

- I had 8 tests fail, 4 warn, 0 skipped and 78 pass

##### {goodpractice} checks

- There are a few comments that could have the line-length shortened to be <80 chars.
Please check with `goodpractice::gp()` and address as necessary and possible.

##### Spell check

`spellcheck::spell_check()` returns several hits.
Please set up spell-checking and specify the language in the DESCRIPTION file.
You can do this with the [{spelling}](https://docs.ropensci.org/spelling/) package.
As a fellow Aussie, I recognise many of these are not misspelled, but {spelling} will allow you to build a wordlist to accept them and not flag in future checks.

```r
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD               FOUND IN
AUD                detailed-example.Rmd:36,37
behaviour          summarising-results-with-predictNMB.Rmd:250
categorised        introduction-to-predictNMB.Rmd:63
centre             introduction-to-predictNMB.Rmd:117
CMD                README.md:8
                   README.Rmd:28
Codecov            README.md:12
                   README.Rmd:30
colour             plot.predictNMBsim.Rd:49,50
Colour             plot.predictNMBsim.Rd:59
coloured           plot.predictNMBscreen.Rd:60
                   plot.predictNMBsim.Rd:37
colours            plot.predictNMBsim.Rd:48
                   summarising-results-with-predictNMB.Rmd:221
FN                 do_nmb_sim.Rd:57
                   creating-nmb-functions.Rmd:24,67
                   introduction-to-predictNMB.Rmd:40,42,63
FNs                creating-nmb-functions.Rmd:121
FP                 do_nmb_sim.Rd:57
                   creating-nmb-functions.Rmd:23,65
                   introduction-to-predictNMB.Rmd:39,42,63
frac               creating-nmb-functions.Rmd:64
grey               summarising-results-with-predictNMB.Rmd:221
inb                make_summary_table.predictNMBscreen.Rd:26,54
                   make_summary_table.predictNMBsim.Rd:25,50
                   make_summary_table.Rd:26,55
                   plot.predictNMBscreen.Rd:38
                   plot.predictNMBsim.Rd:29
INB                introduction-to-predictNMB.Rmd:133
                   summarising-results-with-predictNMB.Rmd:110
magrittr           pipe.Rd:10,12
maximise           creating-nmb-functions.Rmd:32
                   detailed-example.Rmd:32
                   introduction-to-predictNMB.Rmd:25
maximises          do_nmb_sim.Rd:41
                   introduction-to-predictNMB.Rmd:160
maximising         introduction-to-predictNMB.Rmd:194
minimisation       introduction-to-predictNMB.Rmd:123
minimises          do_nmb_sim.Rd:44
minimising         do_nmb_sim.Rd:44
                   screen_simulation_inputs.Rd:53
                   detailed-example.Rmd:95
                   introduction-to-predictNMB.Rmd:143
modelling          creating-nmb-functions.Rmd:35,127,285
nmb                make_summary_table.predictNMBscreen.Rd:26,27,53
                   make_summary_table.predictNMBsim.Rd:25,26,49
                   make_summary_table.Rd:26,27,54
                   plot.predictNMBscreen.Rd:38,39
                   plot.predictNMBsim.Rd:29,30
NMB                do_nmb_sim.Rd:41,56,62,66,98
                   evaluate_cutpoint.Rd:5,16,19,23
                   get_inbuilt_cutpoint.Rd:20
                   get_nmb_sampler.Rd:5,59
                   get_thresholds.Rd:5,14,24
                   make_summary_table.predictNMBscreen.Rd:53,54
                   make_summary_table.predictNMBsim.Rd:49,50
                   make_summary_table.Rd:54,55
                   screen_simulation_inputs.Rd:42,46
                   description:2
                   README.md:21,30,48,73,128,171
                   README.Rmd:37,41,55,75,126,146
                   creating-nmb-functions.Rmd:2,17,19,32,39,60,121,123,150,187,207,208,231,245
                   detailed-example.Rmd:47,68,95,99
                   introduction-to-predictNMB.Rmd:23,25,35,61,65,83,104,117,145,160
                   summarising-results-with-predictNMB.Rmd:85,88,99,110
optimisation       introduction-to-predictNMB.Rmd:123
optimising         do_nmb_sim.Rd:41
                   screen_simulation_inputs.Rd:53
                   detailed-example.Rmd:95,123,178
                   introduction-to-predictNMB.Rmd:143
                   summarising-results-with-predictNMB.Rmd:221
ORCID              predictNMB-package.Rd:23,27,28,33,34
performant         README.md:21
                   README.Rmd:37
pkgcheck           README.md:16
                   README.Rmd:32
predictNMBscreen   print.predictNMBscreen.Rd:5,18
predictNMBsim      print.predictNMBsim.Rd:5,18
PSA                creating-nmb-functions.Rmd:69
QALY               creating-nmb-functions.Rmd:212
QALYs              get_nmb_sampler.Rd:28
                   creating-nmb-functions.Rmd:204,210,212
                   detailed-example.Rmd:32
randomised         detailed-example.Rmd:28
roc                do_nmb_sim.Rd:48
stabilised         summarising-results-with-predictNMB.Rmd:221
summarise          make_summary_table.predictNMBscreen.Rd:26
                   make_summary_table.predictNMBsim.Rd:25
                   make_summary_table.Rd:26
                   plot.predictNMBscreen.Rd:38
                   plot.predictNMBsim.Rd:29
                   summarising-results-with-predictNMB.Rmd:26
summarising        plot.predictNMBscreen.Rd:36
                   introduction-to-predictNMB.Rmd:125
Summarising        README.md:167
                   README.Rmd:144
                   summarising-results-with-predictNMB.Rmd:2
TP                 do_nmb_sim.Rd:57
                   creating-nmb-functions.Rmd:21,64
                   introduction-to-predictNMB.Rmd:37,42
unintuitive        creating-nmb-functions.Rmd:44
visualise          plot.predictNMBscreen.Rd:82
                   plot.predictNMBsim.Rd:79
                   README.md:95
                   README.Rmd:104
                   introduction-to-predictNMB.Rmd:117,133
                   summarising-results-with-predictNMB.Rmd:26,85,88,99,110
visualised         introduction-to-predictNMB.Rmd:117
visualising        README.md:111
                   README.Rmd:112
Whitty             detailed-example.Rmd:180
WTP                creating-nmb-functions.Rmd:204,212
                   detailed-example.Rmd:32
youden             do_nmb_sim.Rd:42
                   detailed-example.Rmd:123
Youden             do_nmb_sim.Rd:42
                   creating-nmb-functions.Rmd:231
                   detailed-example.Rmd:95,97,144,178
                   introduction-to-predictNMB.Rmd:160,194
```

##### Metadata

In the README, there is an incomplete thought on line 79, "cl: (Optional) users can pass a cluster _as the. If it_ is passed, the simulations are run in parallel (faster).", emphasis mine.

Admittedly, this isn't my area of research, but I think that prospective users would benefit from a description of what the fields are in `fx_nmb()` and that `get_nmb_sampler()` returns a function, not an object as this is unusual behaviour in R.
The introductory vignette does cover this, so maybe just a pointer to that section of the document is sufficient for the uninitiated (like me).

There is no "CONTRIBUTING.md", please add one.

##### Other comments

And lastly, RWParsons, could you please add the rOpenSci under review badge to your README?

I'm looking for reviewers now.

***

***
