### 2023-02-06 #566 @RWParsons
Hi @adamhsparks,

This is my first ropensci submission so I'm bound to make mistakes here too!

Thanks for the review - I'll be working on them this week.

A lot of the spell-checks seem to be my Australian spelling. Is it fine for me to leave these as is or should I swap things over to US-English?

I have a `CONTRIBUTING.md` file here (https://github.com/RWParsons/predictNMB/blob/master/.github/CONTRIBUTING.md) Is this in the wrong place?

Thanks!
