+++ pkgdown.yml
@@ -4,9 +4,9 @@ pkgdown_sha: ~
 articles:
   creating-nmb-functions: creating-nmb-functions.html
   detailed-example: detailed-example.html
-  predict_NMB: predict_NMB.html
+  predictNMB: predictNMB.html
   summarising-results-with-predictNMB: summarising-results-with-predictNMB.html
-last_built: 2023-02-06T03:05Z
+last_built: 2023-02-06T03:43Z
 urls:
   reference: https://rwparsons.github.io/predictNMB/reference
   article: https://rwparsons.github.io/predictNMB/articles
