### 2023-02-06 8bb2832
update refs to intro vignette + other README updates
