+++ R/predictNMB-package.R
@@ -4,8 +4,8 @@
 #'
 #' `browseVignettes(package = "predictNMB")`
 #' @keywords internal
-#' @srrstats {G1.0, G1.1} Method to simulate clinical prediction model from AUC - covered by introduction-to-predictNMB vignette.
-#' @srrstats {G2.0a} Differences between do_nmb_sim() and screen_simulation_inputs() in introduction-to-predictNMB vignette.
+#' @srrstats {G1.0, G1.1} Method to simulate clinical prediction model from AUC - covered by the getting started vignette.
+#' @srrstats {G2.0a} Differences between do_nmb_sim() and screen_simulation_inputs() in the getting started vignette.
 #' @srrstats {G1.3} All statistical terminology is defined and explained.
 #' @srrstats {G1.4, G1.4a} Roxygen is used for all documentation.
 #' @srrstats {G1.5, G1.5} No performance claims are made.
