### 2023-02-06 #566 @adamhsparks
Spelling will let you update the wordlist to include any words you know are correct. In the past I have used "EN-AU" but it's occasionally not worked properly for me, so I just update the wordlist and leave the spelling as "EN-US" and use {spelling} to ensure I didn't introduce any new spelling errors after confirming what I wanted for Australian spelling, names, etc.
