### 2023-02-06 8f8c96f
clarify that fx_nmb() is an R function
