+++ DESCRIPTION
@@ -19,7 +19,7 @@ Description: Estimates when and where a model-guided treatment strategy may outp
     a treat-all or treat-none approach by Monte Carlo simulation and evaluation of the Net Monetary Benefit (NMB).
 License: GPL (>= 3)
 Encoding: UTF-8
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.2
 Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
 Imports: 
     assertthat,
