### 2023-02-06 87f9df5
reduce line lengths throughout package code
