+++ R/cutpoint_methods.R
@@ -214,8 +214,9 @@ get_inbuilt_cutpoint <- function(predicted,
 #'
 #'
 #' # get cutpoints using user-defined functions
-#' # These functions must take the \code{predicted} and \code{actual} as arguments
-#' # They can also take \code{nmb} (named vector containing NMB with values for TP, FP, TN, FN)
+#' # These functions must take the \code{predicted} and \code{actual}
+#' # as arguments. They can also take \code{nmb} (named vector containing NMB
+#' # with values for TP, FP, TN, FN).
 #' fx_roc01 <- function(predicted, actual, ...) {
 #'   cutpointr::cutpointr(
 #'     x = predicted, class = actual, method = cutpointr::minimize_metric,
@@ -239,11 +240,18 @@ get_inbuilt_cutpoint <- function(predicted,
 #'   nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
 #' )
 #'
-#' # get a combination of cutpoints from both user-defined functions and inbuilt methods
+#' # get a combination of cutpoints from both user-defined functions and
+#' # inbuilt methods
 #' get_thresholds(
 #'   predicted = runif(1000),
 #'   actual = sample(c(0, 1), size = 1000, replace = TRUE),
-#'   cutpoint_methods = c("fx_roc01", "fx_sum_sens_spec", "youden", "all", "none"),
+#'   cutpoint_methods = c(
+#'     "fx_roc01",
+#'     "fx_sum_sens_spec",
+#'     "youden",
+#'     "all",
+#'     "none"
+#'   ),
 #'   nmb = c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
 #' )
 get_thresholds <- function(predicted, actual, nmb, cutpoint_methods = NULL) {
@@ -267,7 +275,9 @@ get_thresholds <- function(predicted, actual, nmb, cutpoint_methods = NULL) {
   inbuilt_cutpoints <- unlist(inbuilt_cutpoints)
 
   # get cutpoints using user-provided functions
-  non_inbuilt_methods <- cutpoint_methods[!cutpoint_methods %in% inbuilt_methods]
+  non_inbuilt_methods <- cutpoint_methods[
+    !cutpoint_methods %in% inbuilt_methods
+  ]
   if (length(non_inbuilt_methods) != 0) {
     non_inbuilt_cutpoints <- lapply(
       non_inbuilt_methods,
