+++ tests/testthat/test-do_nmb_sim.R
@@ -46,7 +46,8 @@ test_that("do_nmb_sim() results are similar with different seeds", {
   expect_true(all(result_differences < result_tolerances))
 
   thresholds_differences <- abs(
-    colMeans(out1$df_thresholds)[-c(1:3)] - colMeans(out2$df_thresholds)[-c(1:3)]
+    colMeans(out1$df_thresholds)[-c(1:3)] -
+      colMeans(out2$df_thresholds)[-c(1:3)]
   )
   thresholds_tolerances <- abs(colMeans(out1$df_thresholds)[-c(1:3)] * 0.1)
 

[…]
