+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   detailed-example: detailed-example.html
   predictNMB: predictNMB.html
   summarising-results-with-predictNMB: summarising-results-with-predictNMB.html
-last_built: 2023-02-06T04:12Z
+last_built: 2023-02-06T05:59Z
 urls:
   reference: https://rwparsons.github.io/predictNMB/reference
   article: https://rwparsons.github.io/predictNMB/articles
