### 2023-02-07 0de1fb0
avoid warning with new dplyr and update plot snapshots
