+++ NEWS.md
@@ -1,0 +1,11 @@
+# predictNMB (development version)
+
+* Updates to README, vignettes and documentation for clarity
+
+
+# predictNMB 0.0.1
+
+* First submitted version of predictNMB to rOpenSci
+
+* Includes functions for simulating and evaluating clinical prediction models 
+  regarding their the estimated Net Monetary Benefit (NMB)
