+++ .github/workflows/pkgcheck.yaml
@@ -19,5 +19,3 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
-        with:
-          post-to-issue: false
