+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: predictNMB
 Type: Package
 Title: Evaluate Clinical Prediction Models by Net Monetary Benefit
-Version: 0.0.1
+Version: 0.0.1.9000
 Authors@R: c(
     person("Rex", "Parsons", role = c("aut", "cre"),
            email = "rex.parsons94@gmail.com",
