### 2023-02-13 bdfadec
use match.arg() for `what` in plot() and make_summary_table()
