+++ man/predictNMB-package.Rd
@@ -32,7 +32,7 @@ Authors:
 Other contributors:
 \itemize{
   \item Susanna Cramb (\href{https://orcid.org/0000-0001-9041-9531}{ORCID}) [contributor]
-  \item Steve McPhail (\href{https://orcid.org/0000-0002-1463-662X}{ORCID}) [contributor]
+  \item Steven McPhail (\href{https://orcid.org/0000-0002-1463-662X}{ORCID}) [contributor]
 }
 
 }
