+++ DESCRIPTION
@@ -12,7 +12,7 @@ Authors@R: c(
            comment = c(ORCID = "0000-0001-6339-0374")),
     person("Susanna", "Cramb", role = "ctb",
            comment = c(ORCID = "0000-0001-9041-9531")),
-    person("Steve", "McPhail", role = "ctb",
+    person("Steven", "McPhail", role = "ctb",
            comment = c(ORCID = "0000-0002-1463-662X"))
   )
 Description: Estimates when and where a model-guided treatment strategy may outperform 
