### 2023-02-27 c7aee45
Update detailed-example.Rmd
