### 2023-02-28 #566 @emitanaka
# predictNMB

- [x] As the reviewer I confirm that there are no [[conflicts of interest](https://devguide.ropensci.org/policies.html#coi)](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

---

Hi @RWParsons and other authors,

This is a neat package that I imagine will be helpful in decision making for clinical scenarios! The following are points to address some of the missing components from the rOpenSci guideline and some suggestions to hopefully make the package better. Some small typos/grammar fixes were put in as a pull request.

- Contribution guideline appears to be missing (although I note that `ropensci-review-bot` seems to indicate it has found it). All other points under "Community guidelines" are satisfied.

- I recommend calling `plot` as `autoplot` instead. The general convention for `plot` is for base plots and `autoplot` for ggplot outputs. Or alternative suggestion is to make it a synonym: `autoplot.predictNMBsim <- plot.predictNMBsim`.

- I suggest remove `extra_theme` argument in the `plot` and get people to append this to the result of `plot()`.

- I find the argument `return_all_methods` in `get_inbuilt_cutpoint()` a bit awkward and isn't consistent with the main goal of the function (i.e. return a cutpoint). Would it be better to create another function that returns all inbuilt methods? E.g. `get_inbuilt_cutpoint_methods()` that return a vector of method names.

- I suggest the `ci` argument to either be `conf.level` just like in `t.test()` or use more expressive argument

- For `make_summary_table()`, I wonder if you want to leverage the S3 generic `summary()`?

- For `make_summary_table()`, I kind of think you can combine the arguments `what` and `rename_vector`.

Some performance checks for ones I know the correct value:

- Checking `get_sample()` - looks good

```r
# should be roughly 0.7
yardstick::roc_auc(get_sample(0.7, 10000, 0.1), factor(1 - actual), x)
# should be roughly 0.9 and 0.1
table(get_sample(0.7, 10000, 0.1)$actual)/10000
```

- Other functions, I can't think of straightforward tests but I don't see anything particularly wrong in the code.

- Every test and examples seem to be working fine.

I am new to reviewing software so apologies if something in this review is not quite right or missing content I should have included.
