### 2023-02-28 #566 @adamhsparks
Thanks, @emitanaka, for your thoughtful and thorough review.

There is a CONTRIBUTING.md file, it's in the .github directory. I had the same comment. I'm unsure why I missed it or what I thought I was looking for, interesting that you had the same comment though. Also, it does have the proper `URL` and `BugReports` fields in DESCRIPTION.

@Tinula-kariyawasam, are you able to submit your review soon or do you require some more time?
