+++ man/do_nmb_sim.Rd
@@ -71,9 +71,9 @@ expected number of events (\code{sample_size * event_rate}) is met.
 
 \item{min_events}{The minimum number of events to include in the training
 sample. If less than this number are included in sample of size
-\code{sample_size}, additional samples are added until the min_events is met.
+\code{sample_size}, additional samples are added until the \code{min_events} is met.
 The default (\code{NA}) will use the expected value given the
-code{event_rate} and the \code{sample_size}.}
+\code{event_rate} and the \code{sample_size}.}
 
 \item{show_progress}{Logical. Whether to display a progress bar.}
 
