### 2023-03-01 ff4c5c3
use autoplot() instead of plot()
