+++ tests/testthat/test-plot.R
@@ -1,125 +1,125 @@
-### plot.predictNMBsim() tests
+### autoplot.predictNMBsim() tests
 
-test_that("predictNMBsim - plot method (default) works", {
+test_that("predictNMBsim - autoplot method (default) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
 
   # test default plotting of predictNMBsim object
-  predictNMBsim_plot <- plot(obj)
+  predictNMBsim_plot <- autoplot(obj)
   expect_s3_class(predictNMBsim_plot, "gg")
   vdiffr::expect_doppelganger(
-    "plot.predictNMBsim(obj) default",
+    "autoplot.predictNMBsim(obj) default",
     predictNMBsim_plot
   )
 })
 
-test_that("predictNMBsim - plot method (what) works", {
+test_that("predictNMBsim - autoplot method (what) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
 
   # test cutpoints plot from predictNMBsim object
-  predictNMBsim_cutpoints_plot <- plot(obj, what = "cutpoints")
+  predictNMBsim_cutpoints_plot <- autoplot(obj, what = "cutpoints")
   expect_s3_class(predictNMBsim_cutpoints_plot, "gg")
   vdiffr::expect_doppelganger(
-    "plot.predictNMBsim(obj, what = 'cutpoints') default",
+    "autoplot.predictNMBsim(obj, what = 'cutpoints') default",
     predictNMBsim_cutpoints_plot
   )
 
   # test incremental net monetary benefit plot from predictNMBsim object
-  predictNMBsim_inb_plot <- plot(obj, what = "inb", inb_ref_col = "all")
+  predictNMBsim_inb_plot <- autoplot(obj, what = "inb", inb_ref_col = "all")
   expect_s3_class(predictNMBsim_inb_plot, "gg")
   vdiffr::expect_doppelganger(
-    "plot.predictNMBsim(obj, what = 'inb') default",
+    "autoplot.predictNMBsim(obj, what = 'inb') default",
     predictNMBsim_inb_plot
   )
 })
 
-test_that("predictNMBsim - plot method (fill_cols) works", {
+test_that("predictNMBsim - autoplot method (fill_cols) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
 
   # test plotting of predictNMBsim object with new fill colour choices
-  predictNMBsim_christmas_plot <- plot(obj, fill_cols = c("red", "green"))
+  predictNMBsim_christmas_plot <- autoplot(obj, fill_cols = c("red", "green"))
   expect_s3_class(predictNMBsim_christmas_plot, "gg")
   vdiffr::expect_doppelganger(
-    "plot.predictNMBsim(obj, fill_cols = c('red', 'green'))",
+    "autoplot.predictNMBsim(obj, fill_cols = c('red', 'green'))",
     predictNMBsim_christmas_plot
   )
 })
 
 
-### plot.predictNMBscreen() tests
+### autoplot.predictNMBscreen() tests
 
-test_that("predictNMBscreen - plot method (defaults) works", {
+test_that("predictNMBscreen - autoplot method (defaults) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBscreen_object.rds"))
 
   # test default plotting of predictNMBscreen object
-  predictNMBscreen_plot <- plot(obj)
+  predictNMBscreen_plot <- autoplot(obj)
   expect_s3_class(predictNMBscreen_plot, "gg")
   vdiffr::expect_doppelganger(
-    "plot.predictNMBscreen(obj)",
+    "autoplot.predictNMBscreen(obj)",
[…]
