### 2023-03-01 6e5cd02
rename plot files to autoplot
