+++ man/autoplot.predictNMBsim.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/plot.R
+% Please edit documentation in R/autoplot.R
 \name{autoplot.predictNMBsim}
 \alias{autoplot.predictNMBsim}
 \title{Create plots of from predictNMB simulations.}
