+++ man/autoplot.predictNMBscreen.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/plot.R
+% Please edit documentation in R/autoplot.R
 \name{autoplot.predictNMBscreen}
 \alias{autoplot.predictNMBscreen}
 \title{Create plots of from screened predictNMB simulations.}
