+++ vignettes/detailed-example.Rmd
@@ -106,13 +106,13 @@ nmb_simulation
 make_summary_table(nmb_simulation) 
 
 # Demonstrates the range of selected cutpoints under each method
-autoplot(nmb_simulation, what = "cutpoints") 
+autoplot(nmb_simulation, what = "cutpoints") + theme_sim()
 ```
 
 ```{r}
 # Compares the incremental benefit of each alternate strategy to our 
 # current practice (treat all)
-autoplot(nmb_simulation, what = "inb", inb_ref_col = "all") 
+autoplot(nmb_simulation, what = "inb", inb_ref_col = "all") + theme_sim()
 ```
 
 Our first pass shows treating none is likely to be an undesirable option, outperformed by every other method. If we set treat none as the reference strategy, it would have an NMB of \$0. Treat all looks like a pretty good choice, so we could consider making the intervention standard practice. The best option from a NMB perspective looks to be our value optimising or the cost minimising method. The first plot shows that this is actually a lower threshold than the Youden index and other equivalents, which means that we can be a bit less strict in deciding who gets the intervention. So our intervention might be worth using even for some low risk patients.
