+++ NAMESPACE
@@ -16,5 +16,6 @@ export(get_sample)
 export(get_thresholds)
 export(make_summary_table)
 export(screen_simulation_inputs)
+export(theme_sim)
 importFrom(ggplot2,autoplot)
 importFrom(magrittr,"%>%")

[…]
