### 2023-03-01 290476b
create get_inbuilt_cutpoint_methods() and...
small typos in documentation, mainly missing periods.
