+++ man/get_inbuilt_cutpoint_methods.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/cutpoint_methods.R
+\name{get_inbuilt_cutpoint_methods}
+\alias{get_inbuilt_cutpoint_methods}
+\title{Get a vector of all the inbuilt cutpoint methods}
+\usage{
+get_inbuilt_cutpoint_methods()
+}
+\value{
+Returns a vector cutpoint methods that can be used in
+\code{do_nmb_sim()}.
+}
+\description{
+Get a vector of all the inbuilt cutpoint methods
+}
+\examples{
+get_inbuilt_cutpoint_methods()
+}
