### 2023-03-01 8307ae0
update snaps for test-autoplot
