### 2023-03-01 9cef0b7
use summary() instead of make_summary_table()
