+++ tests/testthat/test-make_summary_table.R
@@ -1,19 +1,19 @@
-### make_summary_table.predictNMBsim() tests
-test_that("predictNMBsim - make_summary_table (default) works", {
+### summary.predictNMBsim() tests
+test_that("predictNMBsim - summary (default) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
 
-  # test default make_summary_table for predictNMBsim object
-  predictNMBsim_default_tbl <- make_summary_table(obj)
+  # test default summary for predictNMBsim object
+  predictNMBsim_default_tbl <- summary(obj)
   expect_s3_class(predictNMBsim_default_tbl, "tbl")
   expect_snapshot_output(predictNMBsim_default_tbl)
   expect_true(all(dim(predictNMBsim_default_tbl) == c(8, 3)))
 })
 
-test_that("predictNMBsim - make_summary_table (rename_vector) works", {
+test_that("predictNMBsim - summary (rename_vector) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
 
-  # test that rename_vector works for make_summary_table.predictNMBsim()
-  predictNMBsim_renamed_tbl <- make_summary_table(
+  # test that rename_vector works for summary.predictNMBsim()
+  predictNMBsim_renamed_tbl <- summary(
     obj,
     rename_vector = c("Treat All" = "all", "Treat None" = "none")
   )
@@ -22,11 +22,11 @@ test_that("predictNMBsim - make_summary_table (rename_vector) works", {
   expect_snapshot_output(predictNMBsim_renamed_tbl)
 })
 
-test_that("predictNMBsim - make_summary_table (what) works", {
+test_that("predictNMBsim - summary (what) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
 
-  # test that cutpoints works for make_summary_table.predictNMBsim()
-  predictNMBsim_cutpoints_tbl <- make_summary_table(
+  # test that cutpoints works for summary.predictNMBsim()
+  predictNMBsim_cutpoints_tbl <- summary(
     obj,
     what = "cutpoints"
   )
@@ -35,8 +35,8 @@ test_that("predictNMBsim - make_summary_table (what) works", {
   expect_snapshot_output(predictNMBsim_cutpoints_tbl)
 
   # test that incremental net monetary benefit works for
-  # make_summary_table.predictNMBsim()
-  predictNMBsim_inb_tbl <- make_summary_table(
+  # summary.predictNMBsim()
+  predictNMBsim_inb_tbl <- summary(
     obj,
     what = "inb",
     inb_ref_col = "all"
@@ -47,11 +47,11 @@ test_that("predictNMBsim - make_summary_table (what) works", {
 })
 
 
-test_that("predictNMBsim - make_summary_table (agg_functions) works", {
+test_that("predictNMBsim - summary (agg_functions) works", {
   obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
 
-  # test that agg_functions works for make_summary_table.predictNMBsim()
-  predictNMBsim_custom_fx_tbl <- make_summary_table(
+  # test that agg_functions works for summary.predictNMBsim()
+  predictNMBsim_custom_fx_tbl <- summary(
     obj,
     agg_functions = list(mean = mean, min = min, max = max)
   )
@@ -61,21 +61,21 @@ test_that("predictNMBsim - make_summary_table (agg_functions) works", {
 })
 
 
-### make_summary_table.predictNMBscreen() tests
-test_that("predictNMBscreen - make_summary_table works", {
+### summary.predictNMBscreen() tests
+test_that("predictNMBscreen - summary works", {
[…]
