+++ man/summary.predictNMBscreen.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/make_summary_table.R
+% Please edit documentation in R/summary.R
 \name{summary.predictNMBscreen}
 \alias{summary.predictNMBscreen}
 \title{Create table summaries of \code{predictNMBscreen} objects.}
