### 2023-03-01 d368231
rename test file for get_sample()
