### 2023-03-02 #566 @RWParsons
Thanks, @emitanaka for the thorough review and useful feedback - it's definitely improved the quality of the package and taught me a bit more about best practice!

I have implemented all your suggested changes except for combining the `what` and `rename_vector` for the `make_summary_table()` (now changed to `summary()`). For this method (and same for the `plot()` - now `autoplot()` - method), the `what` pertains to the value being presented (selected cutpoint/NMB/INB) whereas the `rename_vector` corresponds to the cutpoint selection method. Maybe I'm misunderstanding your point and there's a way to combine these that I'm not thinking of though!

On the `extra_theme` comment, I was adding this to the graphic but also `... + theme_bw()`. I have made the change to move the `extra_theme` content to `theme_sim()` and shown how to add this in the documentation and all examples in the README/vignettes. Do you think that I should also avoid applying `theme_bw()` to the graphics being returned (or at least add an argument for controlling this behaviour)?

I do have a performance test for `get_sample()` but it was within a poorly named test file. I have renamed the test file to be `test-get_sample.R` and this assesses the AUC of the resulting model compared to the specified value when calling `get_sample()`.
