+++ tests/testthat/test-do_nmb_sim.R
@@ -12,7 +12,7 @@ test_that("do_nmb_sim() works", {
 
   out <- do_nmb_sim(
     sample_size = 100, n_sims = 10, n_valid = 1000, sim_auc = 0.7,
-    event_rate = 0.1,  fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+    event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
   )
 
   expect_s3_class(out, "predictNMBsim")
