### 2023-03-07 #566 @Tinula-kariyawasam
> Thanks, @emitanaka, for your thoughtful and thorough review.
>
> There is a CONTRIBUTING.md file, it's in the .github directory. I had the same comment. I'm unsure why I missed it or what I thought I was looking for, interesting that you had the same comment though. Also, it does have the proper `URL` and `BugReports` fields in DESCRIPTION.
>
> @Tinula-kariyawasam, are you able to submit your review soon or do you require some more time?

Hi @adamhsparks, just require a bit more time to finish my review. I shall have my review submitted by the 24th March. Thanks for the understanding
