### 2023-03-14 #566 @Tinula-kariyawasam
Hi @RWParsons , here is my review:

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

Congratulations on this great package. It is a really well-thought-out project, and no barring faults or issues are evident at all.

It works really well, is simple and easy to understand.

Thoroughly checks through all edge cases.

Two areas of improvement would entirely satisfy me on this package:

-   The use of more examples and use cases.

-   Efficiency and optimisation of the code, to elaborate;

    -   Try to use vectorisation in your functions as much as possible. For example, instead of using a loop to generate simulated data for **df_nmb_training** and **df_nmb_evaluation**, you could use replicate() and do.call() to generate the data in a more vectorised manner. There are many occurrences of the 'lapply' function being utilised. A specific example is from the **screen_simulation_inputs** where the function used in the simulations loop can be replaced with a vectorised operations (e.g purrr:map(seq_len(nrow(input_grid)), \~ {...})))

    -   Parallelisation: The **lapply()** function can be parallelised using the **furrr** package. By default, **lapply()** runs sequentially, but **furrr** allows us to run the loop in parallel on multiple cores, which can significantly speed up the computation time. We can parallelise the loop using **future_map()** or **future_pmap()**. In the screen_simulation_inputs function, for example **future_map(input_grid, \~ do_nmb_sim(\...))** will run the function **do_nmb_sim()** on each row of **input_grid** in parallel.

    -   Avoiding unnecessary computations: In several places in the code, unnecessary computations are performed. For example, in the **screen_meta** section at the end of screen simulations function, the length of each input parameter is checked to determine if it varies across the screen. Instead of doing this check, we can store a list of the varying parameters at the beginning of the function and use it later to construct the **screen_meta** list. This will avoid unnecessary computations and improve the speed of the function. Another example is from the do_nmb_sim, the **df_nmb_training** and **df_nmb_evaluation** data frames are generated for each simulation, even though they are the same for each simulation. You could generate these data frames once outside the main loop and reuse them for each simulation.

    -   Use of data.table instead of dplyr: If we are dealing with small data sets (in the ranges of hundreds or thousands) then the literature shows that the difference between using data.table and dplyr is minuscule, and it is really a preference of the user. However, to further increase a package's computational speed, I recommend taking advantage of data.table's efficiency. Another note where we can utilise data.table is where there are a few places in the code where the **tidyr::expand_grid()** function is used to create a data frame with all possible combinations of input values. This can be slow for large datasets. Instead, we can use the **CJ()** function from the **data.table** package to create a Cartesian join of the input values. For example, **CJ(sample_size, sim_auc, event_rate, min_events, meet_min_events)** will create a data frame with all possible combinations of **sample_size**, **sim_auc**, **event_rate**, **min_events**, and **meet_min_events**.

    -   The **if (inherits())** statements can be replaced with **if (is.function())** to check if the arguments are functions. This is because **inherits()** checks if an object inherits from a class, which can be slower than checking if an object is a function.

    -   Lastly, Using appropriate data structures can also improve the efficiency of the code. For example, instead of using a data frame to store the results for each simulation, you could use a matrix or an array to reduce memory usage and improve performance.

    -   Many of example are on the do_nmb_sim and screen_simulation_inputs functions are they your two main ones.

-   I have modified and rewritten your code on my own personal device that aligns with the statements I made above then by using the microbenchmark function from the microbenchmark library checked if the run times of the base function or modified had a quicker evaluation based on 1000 runs, at each test the modified was quicker in terms of nanoseconds.

Overall there are no significant changes that I see need to be made to the package, and all tests and examples are working well. Just from a perspective of how we can improve the computational efficiency of this package, some changes need to be made.
