### 2023-03-14 #566 @RWParsons
Thanks for the review and helpful suggestions, @Tinula-kariyawasam !

Where would you like more use-cases and examples? Do you mean in addition to the vignettes or could any particular areas of the vignettes be expanded upon more?

I agree that there's definitely some areas where computational speed could be improved and would be happy to improve things (almost) everywhere you suggest. Could you please share with me the edited version of the package that you have on your device?
