### 2023-03-15 55b9797
[improve efficiency] use glm.fit() rather than glm()
