### 2023-03-15 866d387
[improve efficiency] replace inherits(x, "function") with is.function(x)
