+++ R/autoplot.R
@@ -419,7 +419,7 @@ autoplot.predictNMBscreen <- function(object,
         }
       } else {
         v <- non_x_axis_vars[[i]][[1]]
-        if (inherits(v, "function")) {
+        if (is.function(v)) {
           v <- names(non_x_axis_vars[[i]])[1]
           if (object$pair_nmb_train_and_evaluation_functions &
             x_axis_var %in% c("fx_nmb_training", "fx_nmb_evaluation")) {
