### 2023-03-15 51c1c42
[improve efficiency] update screen_meta generation code
