### 2023-03-15 #566 @RWParsons
Hi @Tinula-kariyawasam, I've just been testing out some of your suggestions and working on some other areas where computation efficiency could be improved now that you've brought my attention there.

> Try to use vectorisation in your functions as much as possible. For example, instead of using a loop to generate simulated data for df_nmb_training and df_nmb_evaluation, you could use replicate() and do.call() to generate the data in a more vectorised manner. There are many occurrences of the 'lapply' function being utilised. A specific example is from the screen_simulation_inputs where the function used in the simulations loop can be replaced with a vectorised operations (e.g purrr:map(seq_len(nrow(input_grid)), ~ {...})))

I know there are a lot of `lapply()`'s about in the code. I just tested swapping the one that's used most heavily (`lapply(seq_len(n_sims), f_iteration_wrapper)` from do_nmb_sim.R) and had almost no performance gain when swapping it for `replicate()`, but did introduce the problem of not being able to pass the iteration number so that the df_nmb_training is indexed appropriately. I think the use of `lapply()` in `screen_simulation_inputs()` are pretty inconsequential since they are only used for manipulating the grid of inputs to `do_nmb_sim()` which is the function that does all the heavy lifting.

> Parallelisation: The lapply() function can be parallelised using the furrr package. By default, lapply() runs sequentially, but furrr allows us to run the loop in parallel on multiple cores, which can significantly speed up the computation time. We can parallelise the loop using future_map() or future_pmap(). In the screen_simulation_inputs function, for example future_map(input_grid, ~ do_nmb_sim(...)) will run the function do_nmb_sim() on each row of input_grid in parallel.

I agree that parallelisation is beneficial here! I had initially done almost exactly as you had suggested - mapping the `input_grid` to `do_nmb_sim()` in parallel. However, I found that this was problematic when the input grid was small and the number of workers available much larger (which I think is likely going to be the case for most users). So, instead, I moved the parallelisation to the simulation-iteration level within `do_nmb_sim()`. I'm not currently using `{furrr}` but using `parallel::parlapply()` instead to keep the dependencies down. I just today tested using furrr::future_map() and had no performance gain compared to parlapply. I acknowledge it may have been a better choice earlier on since I wouldn't have to do the cluster export etc but now that it's set up, I'm reluctant to revert for little performance gain and an additional dependency.

> Avoiding unnecessary computations: In several places in the code, unnecessary computations are performed. For example, in the screen_meta section at the end of screen simulations function, the length of each input parameter is checked to determine if it varies across the screen. Instead of doing this check, we can store a list of the varying parameters at the beginning of the function and use it later to construct the screen_meta list. This will avoid unnecessary computations and improve the speed of the function. Another example is from the do_nmb_sim, the df_nmb_training and df_nmb_evaluation data frames are generated for each simulation, even though they are the same for each simulation. You could generate these data frames once outside the main loop and reuse them for each simulation.

I agree that the code to create the screen_meta object is a bit clunky. I have cleaned this up now so it is a lot cleaner and probably a bit faster although I think this bit of code is likely not affecting total runtime since it's only ever run once.

> Use of data.table instead of dplyr: If we are dealing with small data sets (in the ranges of hundreds or thousands) then the literature shows that the difference between using data.table and dplyr is minuscule, and it is really a preference of the user. However, to further increase a package's computational speed, I recommend taking advantage of data.table's efficiency. Another note where we can utilise data.table is where there are a few places in the code where the tidyr::expand_grid() function is used to create a data frame with all possible combinations of input values. This can be slow for large datasets. Instead, we can use the CJ() function from the data.table package to create a Cartesian join of the input values. For example, CJ(sample_size, sim_auc, event_rate, min_events, meet_min_events) will create a data frame with all possible combinations of sample_size, sim_auc, event_rate, min_events, and meet_min_events.

The only times `expand.grid()` is used is on the simulation input grid, and not on the iteration level of the simulations, so likely only < 50 rows for most users. Also, I intentionally use a tibble and `tidyr::expand_grid()` because it allows me to have a column that contains functions with minimal fuss (the fx_nmb_training and fx_nmb_evaluation). I appreciate the tip for the Cartesian join with data.table but unfortunately I don't think it's going to lead to a large performance gain overall for users due to it's use on only ever small datasets (and never on a per-simulation iteration level).

In terms of other ways to use data.table... currently, almost all the compute time is spent mapping the iteration sequence to `f_iteration_wrapper()`. This is all done either using a `lapply()`, `parlapply()`, or `pblapply()` (if done in parallel). These return a list of outputs and are then transformed back into matrixes and data.frames but that extra transformation is extremely fast comparatively. I'm not sure if there is a way to take advantage of `{data.table}` to speed up this section up. If not, I'm hesitant to add the extra dependency.

> The if (inherits()) statements can be replaced with if (is.function()) to check if the arguments are functions. This is because inherits() checks if an object inherits from a class, which can be slower than checking if an object is a function.

I have updated throughout the package to avoid using `inherits()` for functions and have changed them to `is.function()`. Thanks!

> Lastly, Using appropriate data structures can also improve the efficiency of the code. For example, instead of using a data frame to store the results for each simulation, you could use a matrix or an array to reduce memory usage and improve performance.

Currently, the `df_result` and `df_thresholds` objects are created (when running the simulations) as `matrix` objects and only at the end transformed into data.frames. Today I tested having `get_sample()` return a matrix rather than a `data.frame()` but found no performance benefit. The speed gain in running `get_sample()` was lost due to the additional time required to access columns from the matrix in do_nmb_sim(). (This was surprising to me but I read more about why here - https://stackoverflow.com/questions/54005918/extract-column-from-data-frame-faster-than-from-matrix-why)

> Many of example are on the do_nmb_sim and screen_simulation_inputs functions are they your two main ones.

Correct - these are the two main functions that I expect users to interact with as they abstract away all the efforts of doing the simulation study. `get_nmb_sampler()` may be helpful to users to create the inputs for these functions (fx_nmb_training and evaluation) and all other functions are largely internal.

Another change I made to try to improve performance was swapping out `stats::glm()` for `stats::glm.fit()` and getting the predicted probabilties from the estimated coefficients a bit more directly, rather than using `predict.glm()`. This code is run on each iteration of the simulation and had a 15% speed up on my machine for the model fitting and getting predicted probabilities section within `do_nmb_iteration()`. I appreciate you bringing my attention towards efficiency gains and testing this!
