### 2023-03-16 6126a91
revert glm.fit() change
predict.glm() has some safety features which prevents there being NA predicted values. This was causing issues with the cutpoint calculations. Here, I revert back to using glm() rather than glm.fit() so that I can use predict.glm() rather than doing the matrix multiplication to the coefficients (in an attempt to reduce compute)
