### 2023-03-19 3fa207f
increment roxygen2 version and add ropensci reviewers to DESCRIPTION
