### 2023-03-28 #566 @emitanaka
The fixes look good! My reply to @RWParsons comments before.

> I have implemented all your suggested changes except for combining the `what` and `rename_vector` for the `make_summary_table()` (now changed to `summary()`). For this method (and same for the `plot()` - now `autoplot()` - method), the `what` pertains to the value being presented (selected cutpoint/NMB/INB) whereas the `rename_vector` corresponds to the cutpoint selection method. Maybe I'm misunderstanding your point and there's a way to combine these that I'm not thinking of though!

I think I misunderstood what `rename_vector` is about. I thought initially it was renaming the names under the column `method` or the column name itself (not sure what I was thinking), but I realise now it is not. Do you mind including a working example of how to use `rename_vector`? I just tried with using a character vector but it didn't do anything for me.

> On the `extra_theme` comment, I was adding this to the graphic but also `... + theme_bw()`. I have made the change to move the `extra_theme` content to `theme_sim()` and shown how to add this in the documentation and all examples in the README/vignettes. Do you think that I should also avoid applying `theme_bw()` to the graphics being returned (or at least add an argument for controlling this behaviour)?

I'm mixed with the default application of `theme_bw()`. On one hand, I feel like it should be left as the default theme of ggplot and the responsibility of further modification of the theme should be left to the user. Any change of the "default" theme I feel like should be made in `ggplot2` (hence perhaps not your responsibility). On the other hand, I don't feel the default theme of `ggplot2` is optimal (particularly default colors), so I can understand individual developers modifying the default output. I guess because this is an `autoplot()` and not a new geom, it's understandable that as the developer, you can modify the default theme.

TL;DR: I think it's okay to just apply `theme_bw()` as default on returned object in `autoplo()t`.

> I do have a performance test for `get_sample()` but it was within a poorly named test file. I have renamed the test file to be `test-get_sample.R` and this assesses the AUC of the resulting model compared to the specified value when calling `get_sample()`.

Looks good!
