### 2023-03-29 01c8422
add rename_vector to autoplot and summary examples
