+++ R/summary.R
@@ -74,9 +74,18 @@ get_sim_data <- function(x, what, inb_ref_col = NULL) {
 #'   get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
 #'   sim_screen_obj <- screen_simulation_inputs(
 #'     n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1),
-#'     event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+#'     event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
+#'     cutpoint_methods = c("all", "none", "youden", "value_optimising")
 #'   )
-#'   summary(sim_screen_obj)
+#'   summary(
+#'     sim_screen_obj,
+#'     rename_vector = c(
+#'       "Value_Optimising" = "value_optimising",
+#'       "Treat_None" = "none",
+#'       "Treat_All" = "all",
+#'       "Youden_Index" = "youden"
+#'     )
+#'   )
 #' }
 summary.predictNMBscreen <- function(object,
                                      what = c("nmb", "inb", "cutpoints"),
@@ -161,9 +170,18 @@ summary.predictNMBscreen <- function(object,
 #'   get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
 #'   sim_obj <- do_nmb_sim(
 #'     sample_size = 200, n_sims = 50, n_valid = 10000, sim_auc = 0.7,
-#'     event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+#'     event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
+#'     cutpoint_methods = c("all", "none", "youden", "value_optimising")
 #'   )
-#'   summary(sim_obj)
+#'   summary(
+#'     sim_obj,
+#'     rename_vector = c(
+#'       "Value_Optimising" = "value_optimising",
+#'       "Treat_None" = "none",
+#'       "Treat_All" = "all",
+#'       "Youden_Index" = "youden"
+#'     )
+#'   )
 #' }
 summary.predictNMBsim <- function(object,
                                   what = c("nmb", "inb", "cutpoints"),
