+++ R/autoplot.R
@@ -153,10 +153,19 @@ add_interval <- function(data, conf.level) {
 #'   get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
 #'   sim_obj <- do_nmb_sim(
 #'     sample_size = 200, n_sims = 50, n_valid = 10000, sim_auc = 0.7,
-#'     event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+#'     event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
+#'     cutpoint_methods = c("all", "none", "youden", "value_optimising")
 #'   )
 #'
-#'   autoplot(sim_obj)
+#'   autoplot(
+#'     sim_obj,
+#'     rename_vector = c(
+#'       "Value- Optimising" = "value_optimising",
+#'       "Treat- None" = "none",
+#'       "Treat- All" = "all",
+#'       "Youden Index" = "youden"
+#'     )
+#'   )
 #' }
 autoplot.predictNMBsim <- function(object,
                                    what = c("nmb", "inb", "cutpoints"),
@@ -336,7 +345,8 @@ theme_sim <- function() {
 #'   sim_screen_obj <- screen_simulation_inputs(
 #'     n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1),
 #'     event_rate = c(0.1, 0.2, 0.3),
-#'     fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+#'     fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
+#'     cutpoint_methods = c("all", "none", "youden", "value_optimising")
 #'   )
 #'
 #'   autoplot(sim_screen_obj)
@@ -344,7 +354,13 @@ theme_sim <- function() {
 #'     sim_screen_obj,
 #'     x_axis_var = "event_rate",
 #'     constants = c(sim_auc = 0.8),
-#'     dodge_width = 0.02
+#'     dodge_width = 0.02,
+#'     rename_vector = c(
+#'       "Value-Optimising" = "value_optimising",
+#'       "Treat-None" = "none",
+#'       "Treat-All" = "all",
+#'       "Youden Index" = "youden"
+#'     )
 #'   )
 #' }
 autoplot.predictNMBscreen <- function(object,
