+++ man/summary.predictNMBscreen.Rd
@@ -62,8 +62,17 @@ if (FALSE) {
   get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
   sim_screen_obj <- screen_simulation_inputs(
     n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1),
-    event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+    event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb,
+    cutpoint_methods = c("all", "none", "youden", "value_optimising")
   )
-  summary(sim_screen_obj)
+  summary(
+    sim_screen_obj,
+    rename_vector = c(
+      "Value_Optimising" = "value_optimising",
+      "Treat_None" = "none",
+      "Treat_All" = "all",
+      "Youden_Index" = "youden"
+    )
+  )
 }
 }
