### 2023-03-29 bb1bf51
stop using theme_bw() within autoplot()
and small reorg of autplot.R, add theme_sim() to doc example
