### 2023-03-29 #566 @RWParsons
Thanks, @emitanaka!

> I think I misunderstood what rename_vector is about. I thought initially it was renaming the names under the column method or the column name itself (not sure what I was thinking), but I realise now it is not. Do you mind including a working example of how to use rename_vector? I just tried with using a character vector but it didn't do anything for me.

Sorry for the confusion with the use of `rename_vector`. I've just added the use of it to the examples in the function documentation for autoplot() and summary(). It is otherwise only described/used towards the end of the [summarising results vignette](https://rwparsons.github.io/predictNMB/articles/summarising-results-with-predictNMB.html).

> I'm mixed with the default application of theme_bw(). On one hand, I feel like it should be left as the default theme of ggplot and the responsibility of further modification of the theme should be left to the user. Any change of the "default" theme I feel like should be made in ggplot2 (hence perhaps not your responsibility). On the other hand, I don't feel the default theme of ggplot2 is optimal (particularly default colors), so I can understand individual developers modifying the default output. I guess because this is an autoplot() and not a new geom, it's understandable that as the developer, you can modify the default theme.
> TL;DR: I think it's okay to just apply theme_bw() as default on returned object in autoplo()t.

Right - I can see your POV here and I see the benefit in leaving more options regarding the theme of the plot to the user rather than forcing my preferences on them. I'll leave it to the user to sort out their own themes for the most part.

I've opted to remove the use of `theme_bw()` within the autoplot functions and added some aspects of it to `theme_sim()`. All the examples remain without the use of `theme_bw()` but all cases of using `autplot.predictNMBsim()` include and `+ theme_sim()` (mainly to push the facets close to eachother and box the panels at the top).
