### 2023-03-29 3c8fe30
update summary snaps
snaps of data.frames instead of tibbles due to apparent inconsistent use of shortened names with tibbles(?): https://github.com/RWParsons/predictNMB/actions/runs/4549239318/jobs/8021106189
