+++ tests/testthat/_snaps/summary.md
@@ -69,58 +69,136 @@
 
 # predictNMBscreen - summary works
 
-    # A tibble: 6 x 18
-      sim_auc fx_nmb_train~1 all_m~2 all_9~3 none_~4 none_~5 value~6 value~7 youde~8
-        <dbl> <chr>            <dbl> <chr>     <dbl> <chr>     <dbl> <chr>     <dbl>
-    1     0.7 f1                -1.2 -1.2 t~    -0.4 -0.5 t~   -0.4  -0.5 t~   -0.64
-    2     0.7 f2                -1.2 -1.2 t~    -0.4 -0.5 t~   -0.4  -0.5 t~   -0.64
-    3     0.8 f1                -1.2 -1.2 t~    -0.4 -0.5 t~   -0.4  -0.5 t~   -0.56
-    4     0.8 f2                -1.2 -1.2 t~    -0.4 -0.5 t~   -0.4  -0.5 t~   -0.56
-    5     0.9 f1                -1.2 -1.2 t~    -0.4 -0.5 t~   -0.38 -0.5 t~   -0.47
-    6     0.9 f2                -1.2 -1.2 t~    -0.4 -0.5 t~   -0.38 -0.4 t~   -0.47
-    # ... with 9 more variables: `youden_95% CI` <chr>,
-    #   `cost minimising_median` <dbl>, `cost minimising_95% CI` <chr>,
-    #   `prod sens spec_median` <dbl>, `prod sens spec_95% CI` <chr>,
-    #   roc01_median <dbl>, `roc01_95% CI` <chr>, `index of union_median` <dbl>,
-    #   `index of union_95% CI` <chr>, and abbreviated variable names
-    #   1: fx_nmb_training, 2: all_median, 3: `all_95% CI`, 4: none_median,
-    #   5: `none_95% CI`, 6: `value optimising_median`, ...
+      sim_auc fx_nmb_training all_median   all_95% CI none_median  none_95% CI
+    1     0.7              f1       -1.2 -1.2 to -1.2        -0.4 -0.5 to -0.3
+    2     0.7              f2       -1.2 -1.2 to -1.2        -0.4 -0.5 to -0.3
+    3     0.8              f1       -1.2 -1.2 to -1.2        -0.4 -0.5 to -0.3
+    4     0.8              f2       -1.2 -1.2 to -1.2        -0.4 -0.5 to -0.3
+    5     0.9              f1       -1.2 -1.2 to -1.2        -0.4 -0.5 to -0.3
+    6     0.9              f2       -1.2 -1.2 to -1.2        -0.4 -0.5 to -0.3
+      value optimising_median value optimising_95% CI youden_median youden_95% CI
+    1                   -0.40            -0.5 to -0.3         -0.64  -0.9 to -0.4
+    2                   -0.40            -0.5 to -0.3         -0.64  -0.9 to -0.4
+    3                   -0.40            -0.5 to -0.3         -0.56  -0.8 to -0.4
+    4                   -0.40            -0.5 to -0.3         -0.56  -0.8 to -0.4
+    5                   -0.38            -0.5 to -0.3         -0.47  -0.7 to -0.4
+    6                   -0.38            -0.4 to -0.3         -0.47  -0.6 to -0.4
+      cost minimising_median cost minimising_95% CI prod sens spec_median
+    1                  -0.40           -0.5 to -0.3                 -0.64
+    2                  -0.40           -0.5 to -0.3                 -0.63
+    3                  -0.40           -0.5 to -0.3                 -0.55
+    4                  -0.40           -0.5 to -0.3                 -0.56
+    5                  -0.38           -0.5 to -0.3                 -0.47
+    6                  -0.38           -0.4 to -0.3                 -0.47
[…]
