### 2023-03-30 #566 @RWParsons
Thank you @emitanaka and @Tinula-kariyawasam for the great and constructive reviews - I'm very happy to be able to integrate your feedback and improve the package (and experience this process of code review for the first time).

Since there has been a bit of back and forth with me attending to feedback in previous comments within this isssue, this comment will link back to those and avoid repetiion of the changes that I've made in the package

## Response to @adamhsparks

[Adam's comment](https://github.com/ropensci/software-review/issues/566#issuecomment-1418298705)

Thank you again for the editor check Adam!

My previous response about the contributing file is [here](https://github.com/ropensci/software-review/issues/566#issuecomment-1418423546). The commits where I addressed other points here are listed below:

- Lazy data and R version dependency: [RWParsons/predictNMB `@2352079`](https://github.com/RWParsons/predictNMB/commit/2352079861dad1c04ce3e0f296d5942332451949)
- README and vignettes: [RWParsons/predictNMB `@8bb2832`](https://github.com/RWParsons/predictNMB/commit/8bb2832af54b7d34706d8025aaa4a90fea99e925)
- `goodpractice::gp()` linewidths: [RWParsons/predictNMB `@87f9df5`](https://github.com/RWParsons/predictNMB/commit/87f9df55f8496d6609a79e711b0802515b34fc9c)
- WORDLIST: [RWParsons/predictNMB `@e9a853d`](https://github.com/RWParsons/predictNMB/commit/e9a853d6aabfd73c57effc02e063b05b761a09c0)

## Response to @emitanaka

[Emi's first comment](https://github.com/ropensci/software-review/issues/566#issuecomment-1447725606)

Thanks for the great feedback, Emi! I'm really happy to integrate the changes and use the most appropriate function names/S3 methods as you describe - I definitely learned a lot in the process.

[My response to Emi's first comment](https://github.com/ropensci/software-review/issues/566#issuecomment-1451127891). Associated commits:

- Using `autoplot()` instead of `plot()`: [RWParsons/predictNMB `@ff4c5c3`](https://github.com/RWParsons/predictNMB/commit/ff4c5c382dd27ab1ef46ab3589b178d3ed7e1a29)
- Removing `extra_theme` argument in exchange for `theme_sim()`: [RWParsons/predictNMB `@63176e5`](https://github.com/RWParsons/predictNMB/commit/63176e5c48779ab1f4ba0d9120a0f638ad6811e7)
- Remove `return_all_methods` from `get_inbuilt_methods()` and add `get_inbuilt_cutpoint_methods()`: [RWParsons/predictNMB `@290476b`](https://github.com/RWParsons/predictNMB/commit/290476b0aceb5af8dbe0b684277e7974a9c6fd81)
- Use `conf.level` instead of `ci`: [RWParsons/predictNMB `@16ba525`](https://github.com/RWParsons/predictNMB/commit/16ba5251fed0b1eec0d45ec84b3f692208025e30)
- Use `summary()` instead of `make_summary_table()`: [RWParsons/predictNMB `@9cef0b7`](https://github.com/RWParsons/predictNMB/commit/9cef0b7f21bf87c3daf10c31806ad23c2a770196)

[Emi's second comment](https://github.com/ropensci/software-review/issues/566#issuecomment-1486084094)

Thanks for the opportunity to clarify things regarding the use of `rename_vector` - I added this to the examples in the function documention in [RWParsons/predictNMB `@01c8422`](https://github.com/RWParsons/predictNMB/commit/01c8422b83f54824ea45eb9abcbbdf60c8131041). To address the point about the `ggplot2` themes, I removed the use of `theme_bw()` in [RWParsons/predictNMB `@bb1bf51`](https://github.com/RWParsons/predictNMB/commit/bb1bf5160074f30b679a1724a9fb664a79e81f81).

## Response to @Tinula-kariyawasam

[Tinula's comment](https://github.com/ropensci/software-review/issues/566#issuecomment-1467178633)

Thanks for the feedback Tinula! I'm very happy to clean up the code and give a bit more attention to the areas where efficiency could be improved.

[My response to Tinula's comment](https://github.com/ropensci/software-review/issues/566#issuecomment-1469383543). Associated commits:

- Replace `inherits(x, "function")` with `is.function(x)`: [RWParsons/predictNMB `@866d387`](https://github.com/RWParsons/predictNMB/commit/866d38709898e167afd247b6714873692df3170f)
- Clean up generation of `screen_meta`: [RWParsons/predictNMB `@51c1c42`](https://github.com/RWParsons/predictNMB/commit/51c1c42694bac4ea370b2d7ac7801e1fa2b39f40)
